+++ src/commons/availability/availability-rules/index.js
@@ -7,6 +7,7 @@ const blockPeriodRules = require("./block-period-rules");
 const timePeriodRules = require("./time-period-rules");
 const permissionRules = require("./permission-rules");
 const maxBookingDateRules = require("./max-booking-date-rules");
+const minBookingLeadTimeRules = require("./min-booking-lead-time-rules");
 const { CAPACITY_MODES } = require("./types");
 
 module.exports = {
@@ -19,5 +20,6 @@ module.exports = {
   ...timePeriodRules,
   ...permissionRules,
   ...maxBookingDateRules,
+  ...minBookingLeadTimeRules,
   CAPACITY_MODES,
 };