+++ src/commons/entities/tenant/invitation.js
@@ -8,7 +8,11 @@ class Invitation {
 
     Object.keys(invitationSchemaDefinition).forEach((key) => {
       if (params[key] !== undefined) {
-        this[key] = params[key];
+        if (key === "intendedUserId" && typeof params[key] === "string") {
+          this[key] = params[key].trim().toLowerCase();
+        } else {
+          this[key] = params[key];
+        }
       }
     });
   }